Add a renderpass object to avoid corrupting in GL_CopyFrameBuffer
This fixes the undefined behavior,  where renderpass with loadOp == DONT_CARE
and initialLayout == UNDEFINED was  re-started in GL_CopyFrameBuffer.  As the
framebuffer already was filled with data from before GL_CopyFrameBuffer,
this data was corrupted in vkCmdBeginRenderpass.

To avoid that, another renderpass object is created with loadOp == LOAD and proper
initialLayout is created and used in and after GL_CopyFrameBuffer.

Fixes #24

Authored-by: Lesniewska, Anna <Anna.Lesniewska@intel.com>